+++ CHANGELOG.md
@@ -12,3 +12,9 @@
 ## [0.0.4]
 
 - Fix a bug that may cause `Cannot read property 'scopes' of null` Error.
+
+## [0.0.5]
+
+- Add configurations to set `WParam` and `lParam` for `WM_IME_CONTROL`
+- Add `shiftIm.debug`(*Shift IM: Debug*) command which will help you set Params.
+